+++ src/components/settings/settings-knowledge-base/settings-knowledge-base.js
@@ -2,7 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Field, Form } from 'react-final-form';
 import createFocusDecorator from 'final-form-focus';
-import { FormattedMessage } from 'react-intl';
+import {
+  FormattedMessage,
+  intlShape,
+  injectIntl,
+} from 'react-intl';
 import {
   Headline,
   Icon,
@@ -12,12 +16,14 @@ import {
 
 import SettingsForm from '../settings-form';
 import { KbCredentials } from '../../../constants';
+import NavigationModal from '../../navigation-modal';
 
 const focusOnErrors = createFocusDecorator();
-
-export default class SettingsKnowledgeBase extends Component {
+class SettingsKnowledgeBase extends Component {
   static propTypes = {
     config: KbCredentials.CredentialShape,
+    intl: intlShape,
+    isCreateMode: PropTypes.bool,
     kbCredentials: KbCredentials.KbCredentialsReduxStateShape,
     onSubmit: PropTypes.func.isRequired,
   };
@@ -54,6 +60,22 @@ export default class SettingsKnowledgeBase extends Component {
       }));
     }
 
+    if (kbCredentials.hasSaved) {
+      router.history.push({
+        pathname: `/settings/eholdings/knowledge-base/${config.id}`,
+        state: { eholdings: true, isFreshlySaved: true }
+      });
+
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState(({ toasts }) => ({
+        toasts: [...toasts, {
+          id: `settings-kb-${config.id}`,
+          message: <FormattedMessage id="ui-eholdings.settings.kb.saved" values={{ name: config.attributes.name }} />,
+          type: 'success'
+        }],
+      }));
+    }
+
     if (prevProps.kbCredentials.errors !== kbCredentials.errors) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState(({ toasts }) => ({
@@ -66,11 +88,24 @@ export default class SettingsKnowledgeBase extends Component {
     }
   }
 
+  validateNameField = value => {
+    if (!value) {
+      return <FormattedMessage id="ui-eholdings.validate.errors.settings.kb.name" />;
+    }
+
+    if (value.length > 255) {
+      return <FormattedMessage id="ui-eholdings.validate.errors.settings.kb.name.length" />;
+    }
+
+    return null;
+  }
+
   getInitialValues() {
-    const { config } = this.props;
+    const { config, isCreateMode } = this.props;
 
     const initialValues = {
       rmapiBaseUrl: 'https://sandbox.ebsco.io',
+      name: isCreateMode ? this.getKbCredentialsName() : null,
     };
 
     if (!config) {
@@ -80,10 +115,47 @@ export default class SettingsKnowledgeBase extends Component {
     return Object.assign(initialValues, config.attributes);
   }
 
+  getKbCredentialsName() {
+    const { intl, kbCredentials } = this.props;
+
+    const defaultName = intl.formatMessage({ id: 'ui-eholdings.settings.kb' });
+
+    const kbCredentialsNames = kbCredentials.items.map(({ attributes: { name } }) => name);
+    const defaultNameIsPresent = kbCredentialsNames.some(name => name === defaultName);
+
+    if (!defaultNameIsPresent) {
+      return defaultName;
+    } else {
+      const reg = new RegExp(`${defaultName} \\(\\d*\\)`);
+      const filteredNames = kbCredentialsNames.reduce((acc, name) => {
+        if (name.match(reg)) {
+          const red1 = new RegExp(`${defaultName} \\(`);
+          const number = parseInt(name.replace(red1, ''), 10);
+
+          return [...acc, number];
+        }
+
+        return [...acc];
+      }, []);
+
+      let number = 1;
+
+      while (filteredNames.indexOf(number) !== -1) {
+        number++;
+      }
+
+      return intl.formatMessage(
+        { id: 'ui-eholdings.settings.kb.defaultName' },
+        { number }
+      );
+    }
+  }
+
   render() {
     const {
       onSubmit,
       kbCredentials,
+      isCreateMode,
     } = this.props;
 
     return (
@@ -100,14 +172,25 @@ export default class SettingsKnowledgeBase extends Component {
             title={<FormattedMessage id="ui-eholdings.settings.kb" />}
             toasts={this.state.toasts}
           >
-            <Headline size="xx-large" tag="h3">
-              <FormattedMessage id="ui-eholdings.settings.kb.rmApiCreds" />
-            </Headline>
+            {!isCreateMode && (
+              <Headline size="xx-large" tag="h3">
+                <FormattedMessage id="ui-eholdings.settings.kb.rmApiCreds" />
+              </Headline>
+            )}
 
             {kbCredentials.isLoading ? (
               <Icon icon="spinner-ellipsis" />
             ) : (
               <>
+                <div data-test-eholdings-settings-kb-name>
+                  <Field
+                    name="name"
+                    component={TextField}
+                    label={<FormattedMessage id="ui-eholdings.name" />}
+                    required
+                    validate={this.validateNameField}
+                  />
+                </div>
                 <div data-test-eholdings-settings-kb-url>
                   <Field
                     name="rmapiBaseUrl"
@@ -148,9 +231,19 @@ export default class SettingsKnowledgeBase extends Component {
                 <p><FormattedMessage id="ui-eholdings.settings.kb.url.ebsco.customer.message" /></p>
               </>
             )}
+
+            {isCreateMode && (
+              <NavigationModal
+                label={<FormattedMessage id="ui-eholdings.navModal.areYouSure" />}
+                message={<FormattedMessage id="ui-eholdings.navModal.unsavedChanges" />}
+                when={!formState.pristine}
+              />
+            )}
           </SettingsForm>
         )}
       />
     );
   }
 }
+
+export default injectIntl(SettingsKnowledgeBase);